Allow multiple sources for the same exact library file.
The plugin crashes because it first finds:
/cm/local/apps/cuda-driver/libs/418.67/lib64/libGLESv2.so.2.1.0
And creates a symlink for it in the volume dir.

Then finds it again in:
/usr/lib64/libGLESv2.so.2.1.0

And fails creating the symlink, because the symlink was already there.

Which is installed via libglvnd-gles. Maybe this wasn't the case in the
past, but now it is and NVIDIA docker v1 doesn't like it. Since it's
looking up exact library versions anyway I expect it doesn't even matter
where it links it from, unless there are multiple architectures on the
same system.